Mall of Deals
EMall is an android based application which provides user a better and
new way of searching malls. We have implemented Augmented Reality View
for searching the Malls based on their geo location. There are 3 logins
to the application :
1. User/Shopper
2. Retailer
3. Administrator
Administrator can create an entry for mall. Retailer can Edit their
Discounts and services, and Users can see the discounts.
We have provided various categories of retailers for users convenience.